refactor: document 폴더 구분
