+++ src/main/java/com/sejong/elasticservice/document/consumer/DocumentIndexConsumer.java
@@ -1,7 +1,9 @@
-package com.sejong.elasticservice.document;
+package com.sejong.elasticservice.document.consumer;
 
 import com.sejong.elasticservice.common.name.GroupNames;
 import com.sejong.elasticservice.common.name.TopicNames;
+import com.sejong.elasticservice.document.domain.DocumentIndexEvent;
+import com.sejong.elasticservice.document.repository.DocumentElasticRepository;
 import com.sejong.elasticservice.project.Type;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.annotation.KafkaListener;
