+++ src/main/java/com/sejong/elasticservice/document/domain/DocumentDocument.java
@@ -1,7 +1,8 @@
-package com.sejong.elasticservice.document;
+package com.sejong.elasticservice.document.domain;
 
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import com.sejong.elasticservice.document.DocumentEvent;
 import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.elasticsearch.annotations.*;
