+++ src/main/java/com/sejong/elasticservice/document/domain/DocumentEvent.java
@@ -1,4 +1,4 @@
-package com.sejong.elasticservice.document;
+package com.sejong.elasticservice.document.domain;
 
 
 import lombok.AllArgsConstructor;
