+++ src/main/java/com/sejong/elasticservice/document/repository/DocumentElasticDocumentRepository.java
@@ -1,6 +1,7 @@
-package com.sejong.elasticservice.document;
+package com.sejong.elasticservice.document.repository;
 
 
+import com.sejong.elasticservice.document.domain.DocumentDocument;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 public interface DocumentElasticDocumentRepository extends ElasticsearchRepository<DocumentDocument, String> {
